+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/WebExtensionTest.kt
@@ -225,13 +225,28 @@ class WebExtensionTest : BaseSessionTest() {
             { controller.setAddonManagerDelegate(null) },
             object : WebExtensionController.AddonManagerDelegate {
                 @AssertCalled(count = 3)
+                override fun onEnabling(extension: WebExtension) {}
+
+                @AssertCalled(count = 3)
                 override fun onEnabled(extension: WebExtension) {}
+
+                @AssertCalled(count = 3)
+                override fun onDisabling(extension: WebExtension) {}
 
                 @AssertCalled(count = 3)
                 override fun onDisabled(extension: WebExtension) {}
 
                 @AssertCalled(count = 1)
+                override fun onUninstalling(extension: WebExtension) {}
+
+                @AssertCalled(count = 1)
                 override fun onUninstalled(extension: WebExtension) {}
+
+                @AssertCalled(count = 1)
+                override fun onInstalling(extension: WebExtension) {}
+
+                @AssertCalled(count = 1)
+                override fun onInstalled(extension: WebExtension) {}
             }
         )
 
@@ -2510,28 +2525,46 @@ class WebExtensionTest : BaseSessionTest() {
             }
         })
 
-        val webExtension = sessionRule.waitForResult(
-            controller.install("https://example.org/tests/junit/update-1.xpi")
-        )
-
-        mainSession.reload()
-        sessionRule.waitForPageStop()
-
         sessionRule.addExternalDelegateUntilTestEnd(
             WebExtensionController.AddonManagerDelegate::class,
             { delegate -> controller.setAddonManagerDelegate(delegate) },
             { controller.setAddonManagerDelegate(null) },
             object : WebExtensionController.AddonManagerDelegate {
                 @AssertCalled(count = 0)
+                override fun onEnabling(extension: WebExtension) {}
+
+                @AssertCalled(count = 0)
                 override fun onEnabled(extension: WebExtension) {}
+
+                @AssertCalled(count = 1)
+                override fun onDisabling(extension: WebExtension) {}
 
                 @AssertCalled(count = 1)
                 override fun onDisabled(extension: WebExtension) {}
 
                 @AssertCalled(count = 1)
+                override fun onUninstalling(extension: WebExtension) {}
+
+                @AssertCalled(count = 1)
                 override fun onUninstalled(extension: WebExtension) {}
+
+                // We expect onInstalling/onInstalled to be invoked twice
+                // because we first install the extension and then we update
+                // it, which results in a second install.
+                @AssertCalled(count = 2)
+                override fun onInstalling(extension: WebExtension) {}
+
+                @AssertCalled(count = 2)
+                override fun onInstalled(extension: WebExtension) {}
             }
         )
+
+        val webExtension = sessionRule.waitForResult(
+            controller.install("https://example.org/tests/junit/update-1.xpi")
+        )
+
+        mainSession.reload()
+        sessionRule.waitForPageStop()
 
         val disabledWebExtension = sessionRule.waitForResult(controller.disable(webExtension, source))