+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -20,7 +20,7 @@ exclude: true
   ([bug 1820567]({{bugzilla}}1820567))
 - Add `ViewportMode` annotation to [`viewportMode`][113.7], [`getViewportMode`][113.8] and [`setViewportMode`][113.9].
   ([bug 1820567]({{bugzilla}}1820567))
-- Add [`WebExtensionController.AddonManagerDelegate`][113.10] ([bug 1822763]({{bugzilla}}1822763))
+- Add [`WebExtensionController.AddonManagerDelegate`][113.10] ([bug 1822763]({{bugzilla}}1822763), [bug 1826739]({{bugzilla}}1826739))
 
 [113.1]: {{javadoc_uri}}/GeckoSessionSettings.Builder.html#displayMode(int)
 [113.2]: {{javadoc_uri}}/GeckoSessionSettings.html#getDisplayMode()
@@ -1346,4 +1346,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport(android.content.Context,android.os.Bundle,java.lang.String)
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 04715b0dddf7887089933ed90637737e336fe190
+[api-version]: 3867469003995d18051e031bdc669690295c9171